Bumped Hajk's version number to 2.3.0
A lot has happend in the 2.2.0 release and it's about time to indicate the changes - mainly upgrades of packages. As of 2.3.0 all packages except React and Bootstrap are upgraded to their latest versions.

React is the remaing major upgrade. Regarding Bootstrap, we'll need to discuss it but it's possible we stay at version 3, as 4 dropped support for some older browsers: "Dropped IE8, IE9, and iOS 6 support. v4 is now only IE10+ and iOS 7+. For sites needing either of those, use v3.".