Tests with Edit plugin.
Drawer now on the left and fixed next to collapsed Toolbar. Work remains, mainly on how to display panel when Toolbar isn't visible or when it is expanded (i.e. we can't use the fixed positoning of 72px from left). An idea is to switch to using Flex layout for rendering Toolbar next to Map, instead of inside, as it's rendered now.