Simple snackbar & refactoring of Observers
Added a first, simple implementation of Snackbar (issue #166). All options are not respected yet, but it's good for a start, and we have an API to work against.

Also, tried to bring some clarification into what 'this.observer' is. Now, when a plugin uses the observer pattern, it has a this.localObserver = new Observer() to play with. On the contrary, if a plugin wishes to use the App's (AppModel's) observer, it should be rather called this.globalObserver. So App.js now has this.globalObserver = new Observer(). Hope this will make it easier to see which observer we're subscribing or publishing events to.

Changes are reflected in Dummy plugin.