Added generic, app-wide way to display messages, and more
This commit contains mostly additions releated to issue #166. The Snackbar component that is added is a first step towards closing that issue.

Also, some layout related fixes are added. AppBar did show under map itself, so it was practically invisible, on all Safari browsers (including iPads). This was due to the placing of <header> inside <main>, which was a mistake. That, and some CSS fixes, makes it render correctly in both Chrome and Safari now.

Finally, unneccesary <div>s are being replaced by React.Fragments (with the short syntax of <></>).