Added a new array of "allTools" to buildConfig.json.
The reason is that comments are not allowed in JSON files, hence we can not comment out unused plugins. My intention is to let "allTools" contain all tools, so there's an easy way to see what's available, while "activeTools" only contains plugins that will be build.

In this commit the two arrays are the same so we get a complete build by default.