Bugfixes.
The simple layerswitcher caused unsolved errors. The render method did a remount, this is ok in most cases but the listerner for some events did not work. Group layers must be present in the dom at all time since they have attached event listerners to control the visibility, therefore the rendering method was changed to toggle visibility instead. Also added an event to control set the initial state of the tabs.

The last thing i did was to fix some design errors. The window container is now absolute positioned so it wont "jump" when the scroll is forced.