No more AppBar
This is of course a huge change of how Hajk is constructed, hence a big merge comming up.

I've also made more changes to App.js, that felt natural when the AppBar got removed. The biggest thing is that I've changed where plugins' Windows are rendered.

The focus was on simpification of, both to make it easier for others to work on but also to eliminate things that are not really needed now, when we don't have to handle the AppBar's offset.

There are tons of bugs here and I think there's another huge change coming up, namely HOW we render our Plugins and their Windows. I'm not convinced the current situation (where a button renders Window in Portals) is optimal. I'm inclined to take a look at another approach soon. But this should work for now.

Also, Search is not working currently as I had to change the Input's appearence and didn't focus on functionality just yet.