Completely changed how plugins render their Windows, DrawerButtons and WidgetButtons It's a way better approach to render Windows just once, and let them render whatever button type a given plugin wants to render.
Still a lot of work to do, as for now only LayerSwitcher and Dummy are converted.

I'm thinking of lifting out the main plugin render functionality to a new class, BaseWindowPlugin, and let
all other plugins 'SomePlugin extends BaseWindowPlugin'. I'll try this out soon, but first a commit so I can revert if needed.