Make Lagerhantereran panel title configurable
This change makes the panel title of the layer menu configurable from the map config file , and from admin. There is a new attribute panelTitle which is used for configuration. If there is no panelTitle in the config file or no value for it given, it will default to 'Kartlager' which is the previous hardcoded name.