infoclick - add check for min and max layer resolution.
This commit adds a check for layer resolution within infoclick so that info doesn't show if the layers isn't visible. It is possible to set min and max resolutions for layers from admin/config so set at which scales layers will display but infoclick previously did not take this into account.