Dead functionality removed from documenthandler - see description
Made som changes regarding how we view the getFunctionality-method from search-implemented plugins. Now, we allow the plugin to be used even if getFunctionality returns a null-ish value. We should discuss if this is correct. Maybe we should allow the plugin to be used in the search-component even if getFunctionality is missing entierly?  Current state is that i display a warnig if the getFunctionality-method is missing in the plugin (telling the user that the plugin won't be used in the search), but allowing the method to return null.