Updated how free-solo-searches are performed, see desc
If the user searches by pressing the search-button (or by pressing enter) we make sure to perform a search where all possible combinations are accepted. (As opposed to when a user clicks a autocomplete-entry).

This is done because we want a free-solo-search to give us the same results as shown in the autocomplete-list to not confuse the users.