Big changes to how we handle selected features, see desc
I have reworked how we handle a change of activeFeature versus selecting a feature with the checkbox.

Now, when a user clicks a feature to view it's details we zoom to the feature and highlight it. When the user returns from the details we remove the feature from the selection again.
 - (Before this commit the feature was still selected after returning from the details).

Sorry for the big commit... :)