Bump jsPDF to 2.3.1. (Major version change! Remember to delete node_modules)
This version bump was long overdue, we had some voulnurability warnings (ReDoS was possible via the addImage function in jsPdf).

I've done some initial tests in both the print-plugin as well as in the documentHandler-print. It all seems to work. Would be nice if some other developers tested this before we merge to master.