Changed how margin is added around the image. See desc
The "old way" messed the scale up. This should do the trick!

Now we draw a margin on top of the map-image. This is to make sure that we are not altering the map-image in any way. Not the prettiest solution....

(cherry picked from commit e05182c381900e420ce2fb358aaac500f0aa700a)