Revert "Added 'type: module' to package.json, we want to default to ES6 Modules by default in our code."
There seems to be problems when we build out the project, as all imports get converted to CJS format, which
in turn can't resolve the ESM addition of "import.meta.url". So it's basically this line that causes
trouble, but I don't like the way our builder handles the project (converting all ESM into CJS modules), so
I revert this for now, until we come up with a better solution.

This reverts commit 815dbe81582cb22ddde6a382680dcded5fbd3300.