Closes #938
If the layerType is set as a base-layer in the config-mapper, it should be kept as a base-layer, *even if it has sub-layers*. The old behavior (before this commit) was that the base-layer was "transformed" to a "group-layer" if it had more than one subLayer. Since base-layers might be constructed with several subLayers, we shouldn't do that transformation... If the baseLayer is transformed to a "group-layer" we will get several errors, since the baseLayers does not contain all necessary information to render a "group-layer".