Colors added to TwitterPicker,  reset button for draw & text style
Regarding this issue, two new colors have been added to the color picker for the different styling choices in the drawing tool.
I also created a button that resets the styling of all drawing objects to it's default settings. I personally think that this is  more
user-friendly/a better use case tool than just restoring the color,
Precisely for the reason that this also solves the issue if the user chooses not to want to use the styling values that are stored in local storage when for e.g refreshing Hajk and/or
when the user returns to back to Hajk