update: possibility to change style, buggfix #1177, handle errors
I fixed the whitepage error that prevented buffered objects from being edited due to styling being set to null. As a result, I've also added the ability to change styles on buffered objects through editing, excluding the point size.

Note: There was a slight bug in my previously merged PR regarding issue #1177. When attempting to edit a drawn object such as a "Polygon" or "Circle" with the stroke type set to none, the value was not shown in the strokeSelector dropdown during editing and was automatically set to stroke type "solid" as a result."