Added possibility to buffer from the Sketch plugin (#1460)
* feature: added possibility to buffer from sketch

* update: refactor, added comments & activityid check

Updated the checks for the activityId inside SketchView. Comments added for the new sketch buffer component and some refactoring of the code.

* update: removed unnecessary component, refactor, comment update

* update: comment/s, pluginshown

Since the button in the sketch buffer affects whether the user can draw an object on the map or not I want to ensure that the user has the option to always draw as default when the draw-object plugin is opened.

* update: removed unnecessary dependency

* update: context removed, code refactoring

Decided to remove the context since it's not the best way to handle/set the vector layers in. Decided to refactor and put most of the context code in the parent.

* update: passing props, removed import, variable name changes

* cleanup

* update: folder name, import

Changed the folder name for the newly created buffer components to a more suitable name

* update: enable clicklock

* update: state objects added

The second point of the request is now added and state objects are passed via props

* update: comments

* update: possibility to change style, buggfix #1177, handle errors

I fixed the whitepage error that prevented buffered objects from being edited due to styling being set to null. As a result, I've also added the ability to change styles on buffered objects through editing, excluding the point size.

Note: There was a slight bug in my previously merged PR regarding issue #1177. When attempting to edit a drawn object such as a "Polygon" or "Circle" with the stroke type set to none, the value was not shown in the strokeSelector dropdown during editing and was automatically set to stroke type "solid" as a result."

* update: comment

* update: bufferLayer, comments, styling, passing props

Removed the VectorSource and VectorLayer for the buffer. Updated the comments, removec unnecessary props being passed between components & updated some styling

* update: CHANGELOG.md

---------

Co-authored-by: Jacob Wodzyński <jacob.wod@gmail.com>