feature: clickable infobutton for groups in layerswitcher
A new feature has been added, similar to the one for layers in the layerswitcher, which allows for writing information about a specific layer; it is  now available for groups.

Please note that I also made some adjustments outside the scope of my affected components. Specifically, I utilized and replaced anchor tags with MUI Link for the links in the current group infobutton and also for the links in the layers infobutton