fix: lint errors, queryKey throwing error on service delete
Changed the query hook in the settings.tsx. For some reason when using the handleDeleteService function with the useServiceById hook the page throws an 404 (It can't find the recently deleted service ) even after the delete is successfull. This only happens if the queryKey in the useServiceById is set to "services", when changed to "service" the renderers inside the accordions stop rendering the updated values.

After some research I found a workaround and maybe this is the wrong approach, but until further investigation I'll leave it be as it is.

I used:
const { data: services } = useServices();
 const service = services?.find((s) => s.id === serviceId);