update: included/passing the id prop for LsCheckbox and LsButton, handle layeritem details open and close as mouseEvents in the introduction, more steps and id's added
I got rid of the error that was thrown in the LsCheckBox, forgot to remove the 'slot prop' which is not even a real prop... added some tooltip in the dialog menu when selecting introduction. Disable the layerswitcher introduction button if the layerswitcher is not shown. Passed the children prop to pass on the setLayerDetails globalObserver onClick event that needs a layerId in this case the first index