update: passing appModel to get the layerswitcher title, globalobserver to open the layerswitcher, set some more id's
I made some slight changes to check if the layerswitcher window is closed or open. Instead of checking the display and disable the introductino button i took advantage of the globalobserver to always open the layerswitcher window when the introduction starts. Also added some logic if user is in other tabs other than "Kartlager" automatically switch to this tab before the introduction starts.